Allow exclusions to fixtures recorded
As it stands, all HTTP calls are mocked, including those to
our own service that we are presumably testing! This adds an
option called exclude_scope, which can be an array of
strings of the form array:port to ignore the HTTP calls
of when recording the mocks.

I freely admit to stealing the removeExcludedScopeFromArray
function from poetic/nock-vcr-recorder, where I found this
functionality originally - but nock-vcr-recorder doesn't
catch the BeforeEach tests, which I needed for my project.
It was easier to add this functionality to nockturnal, rather
than vice-versa.